third_party/WebKit/LayoutTests/http/tests/security/powerfulFeatureRestrictions/get-installed-related-apps-on-insecure-origin.html

<!DOCTYPE html>
<title>getInstalledRelatedApps on an insecure origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script>
if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
  window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
} else {
  test(function(test) {
    assert_false('getInstalledRelatedApps' in navigator,
                 'navigator.getInstalledRelatedApps attribute has [SecureContext]');
  }, 'getInstalledRelatedApps requires a secure context');
}
</script>
